Mirroring based on differences in file size and modification time successful
However, if the remote file has a modification time that lies in the same
minute as the local file, the remote file will not be updated, even if
the local file is a few seconds newer (unless the size differs too). This
is because getting the modification time of the remote file is limited
to minute resolution. In the future, I might make an option that says
to upload the local file if the remote file is within one minute, though
this would require some tricky modification of the post-upload missing files
test.