2 bug fixes and made the retry more robust against certain failure
    BUGFIX: in _remove_path_head() (which is used when comparing local and
remote files, and in constructing the list of files to transfer) the path head
was removed anywhere it occurred in the filename, not just at the beginning.
This causes a problem if the user inputs a path relative to e.g. ".", as it will
remove that character from all file names, making it look for the wrong local
file.

    BUGFIX: in _make_remote_dir_if_needed(), when mkdir fails, a second call to
the subprocess's communicate() method resulted in an IOError; this has been
replaced with the output of the previous call to communicate().

    ENHANCEMENT: if the mkdir command on the remote fails, this program no
longer automatically aborts. Instead, it considers that file's sync to have
failed and will try again. This can be overridden with the --no-retry-mkdir (-e)
command line option, which causes the program to exit, like before, if there is
an unexpected error with the mkdir command. This can help debug a problem with a
sync.